+++ commands/serve/web/eHub.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import * as ReactDOM from 'react-dom/client';
 
-import { Root } from './Root';
+import { Root } from './components/Root';
 
 ReactDOM.createRoot(document.querySelector('#app')).render(<Root />);
